Design: Responsive design on Print detail and first layer modal
